Allow SlotMachine to take a hashing function
When two slotMachines have the same number of slots, any string that
hashes to slot 2 in machine 1 will hash to slot 2 in machine 2. In the
current case, this means that mouth 2 will always show up with nose 2.

Adding an option to override the default hashing function allows us to
prevent this issue.