Add configuration to allow fetching of remote IPs
We're currently getting a bunch of garbage (10.x IPs) from `req.ip`,
potentially due to the fact that we're only using the IP and not the
X-Forwarded-For from the request. With this setting, we should be able
to get the "true" IP of the request (if it's there).

See http://expressjs.com/en/api.html#req.ip for more info on the
property and this related setting.