Update test to assert randomness
If we're not asserting that we get a different image on subsequent calls
to the random endpoint, we're not really testing the random endpoint;
the previous tests would pass without the random endpoint existing.

Uses imagemagick's identify function to get a consistent hash of the
image from the inconsistent binary buffer.