Checkpoint: tests are passing
This required a few ts-ignores and I'm sure that a lot of this is still
broken, but the tests are green so we're gonna call it a day.

Next step will be to verify that we can actually run the server, and
that the express middleware still works after the version updates

Then we have to reimplement the build process, and then we can work on
refactoring to use sharp instead of gm.